second cleanup after rebase
