+++ src/main/java/com/bootcamp/demo/restapi/controller/ScooterController.java
@@ -15,7 +15,6 @@ import javax.validation.Valid;
 @RequestMapping(path = "/api/scooters")
 public class ScooterController {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(ScooterController.class);
     private final ScooterService scooterService;
 
     public ScooterController(ScooterService scooterService) {
@@ -25,27 +24,23 @@ public class ScooterController {
     @GetMapping
     @ResponseBody
     public Iterable<Scooter> getScooters() {
-        LOGGER.info("GET SCOOTERS - API endpoint invoked");
         return scooterService.findAllScooters();
     }
 
     @GetMapping("/{scooterId}")
     public Scooter getScooterById(@PathVariable(value = "scooterId") String scooterId) {
-        LOGGER.info("GET SCOOTER BY ID - API endpoint invoked. scooterId = {}", scooterId);
         return scooterService.findScooterById(scooterId);
     }
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
     public Scooter postScooter(@Valid @RequestBody ScooterRequestDTO scooterRequestDTO) {
-        LOGGER.info("POST SCOOTER - API endpoint invoked. Scooter = {}", scooterRequestDTO);
         return scooterService.createScooter(ScooterAssembler.toModel(scooterRequestDTO));
     }
 
     @DeleteMapping("/{scooterId}")
     @ResponseStatus(HttpStatus.ACCEPTED)
     public void deleteScooterById(@PathVariable(value = "scooterId") String scooterId) {
-        LOGGER.info("DELETE SCOOTER BY ID - API endpoint invoked. scooterId = {}", scooterId);
         scooterService.deleteScooterById(scooterId);
     }
 }
