+++ src/main/java/com/bootcamp/demo/service/ScooterService.java
@@ -1,6 +1,7 @@
 package com.bootcamp.demo.service;
 
 import com.bootcamp.demo.model.Scooter;
+import com.bootcamp.demo.service.exception.ItemNotFoundException;
 import com.bootcamp.demo.service.exception.ServiceException;
 import com.google.api.core.ApiFuture;
 import com.google.cloud.firestore.DocumentSnapshot;
@@ -29,7 +30,7 @@ public class ScooterService {
     }
 
     public Scooter findScooterById(String scooterId) {
-        LOGGER.info("FIND SCOOTER BY ID - service function invoked: scooterId = {}", scooterId);
+        LOGGER.info("FIND SCOOTER BY ID - service function invoked. scooterId = {}", scooterId);
 
         try {
             DocumentSnapshot scooter = db.collection(COLLECTION_SCOOTERS_PATH)
@@ -38,18 +39,19 @@ public class ScooterService {
                     .get();
 
             if (!scooter.exists()) {
-                LOGGER.error("FIND SCOOTER BY ID - service function: Scooter does not exists! scooterId = {}", scooterId);
-                throw new IllegalArgumentException();
+                LOGGER.warn("FIND SCOOTER BY ID - Scooter does not exist! scooterId = {}", scooterId);
+                throw new ItemNotFoundException(String.format("Scooter does not exist. scooterId = %s", scooterId));
             }
 
             return scooter.toObject(Scooter.class);
-        } catch (ExecutionException | InterruptedException | IllegalArgumentException e) {
+        } catch (ExecutionException | InterruptedException e) {
             LOGGER.error("FIND SCOOTER BY ID - service function: scooterId = {}.\n Error message: {}.\n StackTrace: {}", scooterId, e.getMessage(), ExceptionUtils.getStackTrace(e));
             throw new ServiceException(e);
         }
     }
 
     public Set<Scooter> findAllScooters() {
+        LOGGER.info("FIND ALL SCOOTER - service function invoked");
         try {
             Iterable<QueryDocumentSnapshot> scooters = db.collection(COLLECTION_SCOOTERS_PATH)
                     .get()
@@ -58,7 +60,7 @@ public class ScooterService {
 
             return StreamSupport
                     .stream(scooters.spliterator(), false)
-                    .map(queryDocumentSnapshot -> queryDocumentSnapshot.toObject(Scooter.class))
+                    .map((queryDocumentSnapshot -> queryDocumentSnapshot.toObject(Scooter.class)))
                     .collect(Collectors.toSet());
         } catch (ExecutionException | InterruptedException e) {
             LOGGER.error("FIND ALL SCOOTER - service function. Error message: {}.\n StackTrace: {}", e.getMessage(), ExceptionUtils.getStackTrace(e));
@@ -67,19 +69,39 @@ public class ScooterService {
 
     }
 
-    public void createScooter(final Scooter scooter) {
-        LOGGER.info("CREATE SCOOTER - service function invoked: scooter = {}", scooter);
+    public Scooter createScooter(final Scooter scooter) {
+        LOGGER.info("CREATE SCOOTER - service function invoked. Scooter = {}", scooter.toString());
 
         try {
             ApiFuture<WriteResult> collectionsApiFuture = db.collection(COLLECTION_SCOOTERS_PATH)
                     .document(scooter.getSerialNumber())
                     .set(scooter);
 
-            String update_time = collectionsApiFuture.get().getUpdateTime().toDate().toString();
+            LOGGER.info("CREATE SCOOTER - created successfully. Update time: {}", collectionsApiFuture.get().getUpdateTime().toDate());
 
-            LOGGER.info("CREATE SCOOTER - created successfully. scooter = {}. Update_time: {}", scooter, update_time);
-        } catch (ExecutionException | InterruptedException e) {
+            return scooter;
+
+        } catch (ExecutionException | InterruptedException | IllegalArgumentException e) {
             LOGGER.error("CREATE SCOOTER - service function: scooter = {}.\n Error message: {}.\n StackTrace: {}", scooter, e.getMessage(), ExceptionUtils.getStackTrace(e));
+            throw new ServiceException(e);
+        }
+    }
+
+    public void deleteScooterById(String scooterId) {
+        LOGGER.info("DELETE SCOOTER - service function invoked. scooterId = {}", scooterId);
+        try {
+            DocumentSnapshot scooter = db.collection(COLLECTION_SCOOTERS_PATH).document(scooterId).get().get();
+
+            if (!scooter.exists()) {
+                LOGGER.warn("DELETE SCOOTER - Scooter does not exist! scooterId = {}", scooterId);
+                throw new ItemNotFoundException(String.format("Scooter does not exist. scooterId = %s", scooterId));
+            }
+
+            ApiFuture<WriteResult> writeResult = db.collection(COLLECTION_SCOOTERS_PATH).document(scooterId).delete();
+
+            LOGGER.info("DELETE SCOOTER - service function. Update time : {} | scooterId = {}", writeResult.get().getUpdateTime(), scooterId);
+        } catch (ExecutionException | InterruptedException e) {
+            LOGGER.error("DELETE SCOOTER - service function. scooterId = {}.\n Error message: {}. \n StackTrace: {}", scooterId, e.getMessage(), ExceptionUtils.getStackTrace(e));
             throw new ServiceException(e);
         }
     }
