+++ src/main/java/com/bootcamp/demo/service/exception/ItemNotFoundException.java
@@ -2,10 +2,6 @@ package com.bootcamp.demo.service.exception;
 
 public class ItemNotFoundException extends RuntimeException {
 
-    public ItemNotFoundException() {
-        super("");
-    }
-
     public ItemNotFoundException(String message) {
         super(message);
     }
